LLM: refactor(tracing): move TraceProvider to separate module and replace global with accessor
- Move TraceProvider and related logic from setup.py to new provider.py module
- Replace direct GLOBAL_TRACE_PROVIDER references with get_trace_provider() accessor throughout tracing code
- Update init and util imports to use get_trace_provider
- Call set_trace_provider(TraceProvider()) on init
- Remove ID and time generation logic from util, delegate to TraceProvider
- Update SpanImpl to always use passed-in span_id
