+++ src/agents/tracing/util.py
@@ -1,21 +1,21 @@
-from .setup import GLOBAL_TRACE_PROVIDER
+from .setup import get_trace_provider
 
 
 def time_iso() -> str:
     """Return the current time in ISO 8601 format."""
-    return GLOBAL_TRACE_PROVIDER.time_iso()
+    return get_trace_provider().time_iso()
 
 
 def gen_trace_id() -> str:
     """Generate a new trace ID."""
-    return GLOBAL_TRACE_PROVIDER.gen_trace_id()
+    return get_trace_provider().gen_trace_id()
 
 
 def gen_span_id() -> str:
     """Generate a new span ID."""
-    return GLOBAL_TRACE_PROVIDER.gen_span_id()
+    return get_trace_provider().gen_span_id()
 
 
 def gen_group_id() -> str:
     """Generate a new group ID."""
-    return GLOBAL_TRACE_PROVIDER.gen_group_id()
+    return get_trace_provider().gen_group_id()
