+++ tests/conftest.py
@@ -7,7 +7,7 @@ from agents.models.openai_chatcompletions import OpenAIChatCompletionsModel
 from agents.models.openai_responses import OpenAIResponsesModel
 from agents.run import set_default_runner
 from agents.tracing import set_trace_processors
-from agents.tracing.setup import GLOBAL_TRACE_PROVIDER
+from agents.tracing.setup import get_trace_provider
 
 from .testing_processor import SPAN_PROCESSOR_TESTING
 
@@ -43,7 +43,7 @@ def clear_default_runner():
 @pytest.fixture(autouse=True, scope="session")
 def shutdown_trace_provider():
     yield
-    GLOBAL_TRACE_PROVIDER.shutdown()
+    get_trace_provider().shutdown()
 
 
 @pytest.fixture(autouse=True)
